requirements.txt
django==4.2.16
djangorestframework==3.15.1
requests==2.32.3
django-cors-headers==4.3.1
django-redis==5.4.0
python-dotenv==1.0.1
structlog==24.4.0
redis>=4.5.0